repo_src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
psycopg2-binary # Keep if you plan to support PostgreSQL, otherwise remove for pure SQLite
google-generativeai>=0.3.0 